# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/timeoff_types:
    get:
      tags:
        - TimeoffTypes
      summary: List TimeoffTypes
      description: Get a paginated list of TimeoffTypes with optional filtering and sorting
      operationId: listTimeoffTypes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeoffTypeList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TimeoffTypes
      summary: Create TimeoffType
      description: Create a new TimeoffType
      operationId: createTimeoffType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeoffTypeRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeoffType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    TimeoffType:
      type: object
      required:
        - id
        - company_id
        - name
        - is_paid
        - allow_carry_forward
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        code:
          nullable: true
          type: string
          maxLength: 40
        is_paid:
          type: boolean
        allow_carry_forward:
          type: boolean
        metadata:
          type: object
...
    CreateTimeoffTypeRequest:
      type: object
      required:
        - company_id
        - name
        - is_paid
        - allow_carry_forward
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        code:
          nullable: true
          type: string
          maxLength: 40
        is_paid:
          type: boolean
        allow_carry_forward:
          type: boolean
        metadata:
          type: object
...
    TimeoffTypeList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TimeoffType'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []