# src/presentation/http/timeoff_type_handler.rs
//! TimeoffType REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{TimeoffTypeService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateTimeoffTypeDto, UpdateTimeoffTypeDto, PatchTimeoffTypeDto, TimeoffTypeResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum TimeoffTypeError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for TimeoffTypeError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for TimeoffTypeError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "TIMEOFFTYPE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "TIMEOFFTYPE_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "TIMEOFFTYPE_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "TIMEOFFTYPE_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for TimeoffType.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /timeoff_types | List with pagination |
/// | POST | /timeoff_types | Create new |
/// | GET | /timeoff_types/:id | Get by ID |
/// | PUT | /timeoff_types/:id | Full update |
/// | PATCH | /timeoff_types/:id | Partial update |
/// | DELETE | /timeoff_types/:id | Soft delete |
/// | POST | /timeoff_types/bulk | Bulk create |
/// | POST | /timeoff_types/upsert | Upsert |
/// | GET | /timeoff_types/trash | List deleted |
/// | POST | /timeoff_types/:id/restore | Restore |
/// | DELETE | /timeoff_types/empty | Empty trash |
/// | GET | /timeoff_types/:id/deleted | Get deleted by ID |
/// | DELETE | /timeoff_types/trash/:id | Permanent delete from trash |
/// | GET | /timeoff_types/count | Count active entities |
/// | GET | /timeoff_types/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(TimeoffTypeService::with_repository(repository));
/// let router = create_timeoff_type_routes(service);
/// ```
pub fn create_timeoff_type_routes(service: Arc<TimeoffTypeService>) -> Router {
    BackboneCrudHandler::<TimeoffTypeService, TimeoffType, CreateTimeoffTypeDto, UpdateTimeoffTypeDto, TimeoffTypeResponseDto>::routes(
        service,
        "/timeoff_types",
    )
}

/// Create Axum router with only the read (GET) endpoints for TimeoffType.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_timeoff_type_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_timeoff_type_read_routes(service: Arc<TimeoffTypeService>) -> Router {
    BackboneCrudHandler::<TimeoffTypeService, TimeoffType, CreateTimeoffTypeDto, UpdateTimeoffTypeDto, TimeoffTypeResponseDto>::read_routes(
        service,
        "/timeoff_types",
    )
}

/// Create Axum router with only the write (mutation) endpoints for TimeoffType.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_timeoff_type_write_routes(service: Arc<TimeoffTypeService>) -> Router {
    BackboneCrudHandler::<TimeoffTypeService, TimeoffType, CreateTimeoffTypeDto, UpdateTimeoffTypeDto, TimeoffTypeResponseDto>::write_routes(
        service,
        "/timeoff_types",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_timeoff_type_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<TimeoffTypeService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_timeoff_type_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}